Make sed line FreeBSD compatible
The recent addition of help.h to the build uses a sed command that is
apparently incompatible with FreeBSD's sed.  It pukes with complaints
about the use of "$$" and the a command.

Since the entire purpose of this part of the sed command is to add a
semicolon at the end of the help string, I have replaced it with a
simple echo of ";" appended to help.h.