Remove src/help.h from git tracking
src/help.h is auto generated by the makefile, and is ignored by
.gitignore, but is still present in the repo.

Since it is regenerated automatically from src/help.txt there is no
need for it in the repo.  Removing it also makes it so that when one
does a "make clean" and help.h is deleted, git doesn't think the repo
is dirty.