add option to allocate pseudotty for subprocess
Added a `-pty` option to allocate a pseudo-terminal
via `posix_openpt()` and associated interfaces, and
connect to it as the controlling terminal when
setting up the process that that `tek4010` creates
and runs.  This allows us to run e.g. a shell that
expects access to a (pseudo)tty device under `tek4010`,
or attach to an instance of GNU screen or something
similar, without having to use `rsh` or `ssh -tt`
first.